revised how the sandbox should be
