--- sandbox/App/Console/MySandboxCommand.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Sandbox\App\Console;
-
-use Bootstrap\Console\SlayerCommand;
-
-class MySandboxCommand extends SlayerCommand
-{
-    protected $name = 'sandbox:test';
-
-    protected $description = 'This is just a sandbox command';
-
-    public function slash()
-    {
-
-    }
-}
