--- sandbox/App/routes.php
@@ -1,9 +1,0 @@
-<?php
-
-use Bootstrap\Facades\Route;
-
-Route::add('/sandbox/test', [
-    'module'     => 'sandbox',
-    'controller' => 'Sandbox',
-    'action'     => 'test',
-]);
