+++ sandbox/acme/src/app/routes.php
@@ -1,0 +1,9 @@
+<?php
+
+use Bootstrap\Facades\Route;
+
+Route::add('/acme/test', [
+    'module'     => 'acme',
+    'controller' => 'Acme',
+    'action'     => 'test',
+]);
