+++ sandbox/acme/src/resources/views/test.volt
@@ -3,7 +3,7 @@
         <title>Tests</title>
     </head>
     <body>
-        This is a sandbox tests.<br>
+        This is a sandbox tests using Acme/Acme.<br>
 
         {{ lang.get('message.sample') }}
     </body>
